LFOInput tidy with shared helper function
